requirements.txt
fire==0.7.0
google-cloud-aiplatform[adk,agent_engines]
nest-asyncio